Allow ConcurrentMultiPartUploadsMockFsRepository to propagate IOExceptions (elastic/elasticsearch-serverless#4082)
This repository uses an Executor to perform chunk copies concurrently.
If a copy task generates an IOException then the Executor wraps it in
an ExecutionException, which in turn is converted into an AssertionError
by safeGet. This causes injected corruption errors that would otherwise
be handled by the VBCC upload task to escape, failing the test.

To make concurrent upload have similar IOException behaviour to sequential
upload, this change unwraps an IOException found inside an ExecutionException
inside an AssertionError if one is present.

Fixes elastic/elasticsearch-serverless#3962