Do not cache potentially stale primary term read in StatelessReshardIT (elastic/elasticsearch-serverless#4189)
The failure being addressed is that the final assertion in the test
expects a lower primary term on the source than what is actually
present. This appears to be because we cache the source primary term
we read from cluster state right after the source has recovered,
and we may be viewing an intermediate cluster state that is one
primary term stale.

Reading the primary term of the source directly at the end of the
test should make it harder to hit such a race.