Only monitor acquired searcher time for can_match and search source values (elastic/elasticsearch-serverless#4243)
We acquire searches for many reasons in Elasticsearch. Some are for reading realtime values from the translog, for reading stats, and for actually executing search requests.

For prefetching newly created commits we want to only do so for indices that have been recently searched (within a configurable idle time).

Before this PR none of the indices were never idle (so we were always prefetching) because we considered all acquired searchers as searchers used to execute search request. This changes the behaviour to only consider the searchers that are created for the can_match phase or search (ES|QL uses the `search` type searcher too).

This has the effect of now only prefetching new commits for indices that have been searched within the configured idle time (default to 5 minutes).