Track aquired Lucene index commits across engine resets (elastic/elasticsearch-serverless#4281)
This solution is based on LocalCommitsRefs which tracks the acquired commit generation at the shard level.
This instance is then passed to the Lucene index deletion policies used by the IndexEngine and HollowIndexEngine.

It only take care of commits acquired using Engine#acquireIndexCommit(boolean) for snapshots.

Relates ES-10929
Co-authored-by: Tanguy Leroux tlrx.dev@gmail.com

This PR re-introduces elastic/elasticsearch-serverless#4140 with the following enhancements:

- Add setting to enable/disable stateless custom deletion policy
- Keep track of acquired commits for the commit listener (just for assertions) and some simplifications