Avoid prefetching historical data (elastic/elasticsearch-serverless#4317)
With commit prefetching we aim to prefetch data that's being ingested for
indices that are not idle (that are being searched).

We noticed that after long idle times a search operation could trigger the
prefetching of all the previous commit files, not just the latest generation.

This updates the search commit prefetching logic to:
1. when skipping prefetching due to idleness mark the new commit notifcations as
"prefetched" so they're not downloaded anymore in the future when a new commit
notification comes in
2. move the prefetched marker at the beginning of the current term and generation
(so only the current term and generation are prefetched) upon the receipt of the
first commit notification.