[Test] Ensure file-based settings processed and disable file deletion (elastic/elasticsearch-serverless#4312)
There were 2 issues with the test failure in elastic/elasticsearch-serverless#4302: 1. settings.json was
not updated after removing the project config files. This is already
fixed in elastic/elasticsearch-serverless#4295 2. Not waiting for settings.json to be processed by
file-based settings. This PR fixes it.

Both issues can sometimes lead to mixed processing between previous
deletion and recreation or vice versa.

The issue with failure in elastic/elasticsearch-serverless#4038 is because File-based settings service
does not support file deletion and no clean up for project deletion.
This PR fixes it by temporary commentting out the deletions for project
config files and add a TODO with an existing issue linked.

Resolves: elastic/elasticsearch-serverless#4302 Resolves: elastic/elasticsearch-serverless#4308 Resolves: elastic/elasticsearch-serverless#4314 Resolves: elastic/elasticsearch-serverless#4330