Don't fallback to the blob store if the index node has crashed in GetVBCCChunk (elastic/elasticsearch-serverless#4359)
TransportGetVirtualBatchedCompoundCommitChunkAction throws a ResourceNotFoundException
if the preferred indexing node is not in the cluster state.

If the node departed gracefully and we cannot find it anymore in the cluster state it means that the chunk
that we're trying to read must be uploaded into the blob store, but somehow the reader is a bit behind.
In that case it's safe to just throw ResourceNotFoundException.

On the other hand, if the primary shard is unassigned, it means that the primary node crashed and
it's likely that the BCC wasn't uploaded and the blob store would return a confusing 404.
In such cases, we just fail right away.

Relates elastic/elasticsearch-serverless#4343