Fix dynamic settings plumbing in SearchCommitPrefetcher (elastic/elasticsearch-serverless#4424)
The SearchCommitPrefetcher gets created a bit later in a node lifecycle
and we have one instance per index.
Before this change, we'd initialize and watch the dynamic settings in the
prefetcher constructor (a pattern widely adopted within Elasticsearch)
however, previous usages initialize the settings earlier in the node lifecycle.
`initializeAndWatch` is trappy as it reads the `settings` that are encapsulated
in the `ClusterSettings` when intializing but NOT the `lastAppliedSettings`
(we should address this in another workstream). This could lead to `initializeAndWatch`
to miss reading an updated dynamic cluster setting (that now lives in `lastAppliedSettings`)
because `SearchCommitPrefetcher` is initialized late in the node lifecycle.

This commit separates the dynamic settings outside the `SearchCommitPrefetcher` and
initializes them earlier in the node lifecycle (like the other Elasticsearch usages).

This also avoids creating many unnecessary settings listener instances.