Fix testSearchNodeDoesNotFallbackToBlobStoreOnIndexNodeCrash (elastic/elasticsearch-serverless#4441)
Wait for search node to see the primary shard become unassigned
before letting the request for the VBCC chunk through.

This test is meant to check that if the index shard has crashed, then
requests for VBCC chunks will fail immediately instead of falling back to
the object store where they may produce upsetting 404 errors. But there
is a window between the index node crashing and the search node realizing
it during which it may fall back to the object store anyway, failing
the assert in the test. To work around this, the test now explicitly
waits for the shard to become unassigned before letting the VBCC chunk
request through.

This only makes the test deterministic - the underlying race is in the
production code, which means that we are only best effort avoiding the
object store fallback and could still sometimes get 404s on this path,
just less frequently than before. I haven't thought of a cheap way to
avoid that so far.

Fixes elastic/elasticsearch-serverless#4405