Reduce prewarming thread count (elastic/elasticsearch-serverless#4449)
We sometimes see the prewarming threads burning a bunch of CPU (more
than 35%) particularly when migrating lots of shards after a scaling
event. Todays we allow up to 4*#CPUs of them to run at once, which will
max out the node's CPU allowance and get in the way of real work.

This commit drops the allowance to 2*#CPUs to avoid this CPU saturation
effect.