Prohibit top level _project mappings in indices and runtime fields in serverless (elastic/elasticsearch-serverless#4596)
In order to prepare for cross-project search where project tags (defined in cloud) can be 
used in search and ES|QL queries, we need to prohibit any Lucene or Elasticsearch runtime
fields that start with _project, as that will be the prefix by which we identify project metadata
tags in queries (stored in cluster state rather than a Lucene index).

This PR leverages the RootObjectNamespaceValidator framework added to elasticsearch
core repo in https://github.com/elastic/elasticsearch/pull/132360 by creating a 
`ServerlessRootObjectMapperNamespaceValidator` that prohibits any top level field 
(or "namespace") starting with `_project` both in standard index mappings and in runtime mappings.