Add `ThreadWatchdog` to `ClusterApplierService` (elastic/elasticsearch-serverless#4636)
Adds another layer of detection of slow activity on the cluster applier
thread. In particular this can detect activity that isn't included in an
`UpdateTask`, which particularly may include completing an expensive
listener attached to a `ClusterStateObserver`. Moreover it captures a
thread dump if slow activity is detected.

Serverless counterpart of elastic/elasticsearch#134361