Custom IDs and more logging for hollow stress IT (elastic/elasticsearch-serverless#4721)
Test issue elastic/elasticsearch-serverless#4715 signifies an impossible state: during an update,
it gets a higher seqno in the pre-update for the doc than the one
it finds ultimately when it tries to insert the updated doc.

It may be that there is some rare chance of racing with the
indexing threads that generate random IDs. So using explicit
docIds that atomically increment to avoid any chances.

Also, enable more logging, which also logs doc IDs as they
are being indexed/updated, so we can more easily debug this
in case the test is muted again. We should disable the logging
if we see it is not occurring again.

Relates elastic/elasticsearch-serverless#4715