Skip explain msg in `StatelessAllocationDecider` (elastic/elasticsearch-serverless#4746)
Calling `statelessNodeRole(roles)` before `allocation.decision(...)`
means that we construct this string always and then immediately discard
it (unless running in explain mode). Stream-based computations are
remarkably expensive, and even if we refactored it not to use streams
it'd still at least have to allocate the string as well as various other
bits and pieces. With this commit we call `allocation.debugDecision`
first and avoid any allocation on the hot path.