Use search request shard count summary on search shard when opening reader (elastic/elasticsearch-serverless#4871)
During resharding, a search shard may contain documents that belong to a sibling shard for a period of time after that shard is first created. If a search is querying both shards, then each shard should filter out the documents that now belong to the other. But if the request is only querying the original shard during this period, because the new shard wasn't ready when it created the request, then the original shard should not filter.

The request has recently been updated to include a summary of the shard count used when the request was created. This change consumes that to make the correct filtering decision.