Track and use tier-wide average task execution time during scaling (elastic/elasticsearch-serverless#4891)
With this change the master tracks a tier-wide average task execution
time for the Write threadpool and uses it to re-estimated the queued
work for the WRITE threadpool. Each node reports a stable average task
execution time (if it has one) along with its ingestion load. The
reported value is for all the different write threadpools, although we
only use the one for WRITE. A node considers its latest
`averageTaskExecutionEWMA` in the executor stats stable if:

- `INITIAL_INTERVAL_TO_CONSIDER_NODE_AVG_TASK_EXEC_TIME_UNSTABLE` has passed since the node's first shard assignment (essentially since the node has started, but to make this more accurate).
- If a scaling event is in progress (based on existence of shutdown markers), `INITIAL_SCALING_WINDOW_TO_CONSIDER_AVG_TASK_EXEC_TIMES_UNSTABLE` has passed from the start of the scaling event.

Both of these are by default set to 5 minutes. The idea is to NOT rely
on the actual `averageTaskExecutionEWMA` within the first 5 min for a
new node, and within the first 5 min of a scaling event for the entire
tier. There are cases where these two overlap. 

The reason is that during these periods there are often very
high/unpredictable jumps in the average task execution time for the
WRITE threadpool which causes very high estimates for the queued work.
I've seen this jump for both new nodes and existing nodes during
scaling. While both of these need to be potentially investigated
separately, Using a tier-wide value can make these estimates more
predictable and still take into account the actual queue size, although
re-estimated using a tier-wide average. If a tier-wide average is not
available (we don't consider zero a valid value to use in the average),
we fall back to using what the node has reported and do not re-estimate
the queued work.

I understand that there are different possible variations of what I'm
doing here, but my aim is to get a basic/simple version out, disabled by
default, try it out and gradually improve it if necessary. 

~The code is not entirely polished, and would need more testing and the
existing tests need a bit more detail. I'll get to that once we agree on
the high-level implementation here.~

Relates https://github.com/elastic/elasticsearch-serverless/pull/4826,
https://github.com/elastic/elasticsearch-serverless/pull/4836,
https://github.com/elastic/elasticsearch-serverless/pull/4849 Resolves
ES-13462