Refactor method that consolidates iteration over referenced BCC and CC (elastic/elasticsearch-serverless#4982)
The goal of this refactoring is to introduce a new method: ObjectStoreService#readReferencedCompoundCommits.
This method consolidates the iteration over referenced BCCs together with the iteration over the containing CCs of those BCCs (this logic was previously divided between the readBlobsAndComputeBlobFileRanges and computeBlobFileRanges methods). The new method only reads and iterates blobs, it does not compute blob file ranges (which is left for the argument listener to do next).
It will be used to iterate and pre-warm referenced BCC/CC on the search shards recovery code path.