Create a structure for ReplicasLoadBalancingScaler (elastic/elasticsearch-serverless#5034)
This adds a new component, ReplicasLoadBalancingScaler with a structure proposal
for how it might work and where it'll be used (i.e. in `ReplicasUpdaterService`).

As a first step it calls the indices stats API to collect the indices search load
and computes the indices relative search load.
(once we have the desired topology in place we'll be able to compute the replicas
recommendations and call the service from `ReplicasUpdatedService` - this is just
an initial step that'll also enable working in parallel on the feature).

Additionally, this PR also adds a service to enable/disable the replicas for
load balancing and a feature flag.