Create POST autoscaling metrics API (elastic/elasticsearch-serverless#5027)
Creates an endpoint:
```
POST /_internal/serverless/autoscaling {
 "desired": { 
 "tier": {
 "search": {
 "replicas": 2,
 "memory": "8Gi",
 "storageRatio": 32.5,
 "cpuRatio": 0.5,
 "cpuLimitRatio": 1.5
 },
 "index": { … }
 } 
 } 
}
```
Alongside the existing `GET /_internal/serverless/autoscaling`. The functionality is the same (it returns metrics) but the POST kicks off a transport action which stores the body in a singleton on the master node: DesiredTopologyContext. This will eventually be used in the ReplicasUpdaterService.