Refactor ObjectStoreService read*UsingCache methods (elastic/elasticsearch-serverless#5099)
For the most part, this refactoring PR changes the signature of the ObjectStoreService#read*UsingCache methods to take in a BlobStoreCacheDirectory parameter rather than a IndexBlobStoreCacheDirectory one. This is done in order to allow reusing those method with an argument of the SearchDirectory type (which is a subclass of the BlobStoreCacheDirectory class (but not of the IndexBlobStoreCacheDirectory class)).